podpięcie api do tras, koniec z hardcodowanymi danymi

+++ mySzlak/app/src/main/java/com/example/myszlak/Trail.kt
@@ -10,7 +10,7 @@ data class Trail(
     val id: Int,
     val name: String,
     val description: String,
-    val length: Float,
+    val length: Double,
     val type: TrailType,
     val imageId: String
 )
